pyproject.toml
[tool.poetry]
name = "remarkable-obsidian-sync"
version = "0.1.0"
description = ""
authors = ["Florian Labarre <contact@florianlabarre.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
pdf2image = "^1.16.3"
rmscene = {path = "../rmscene"}
python-dotenv = "^1.0.0"
psutil = "^5.9.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
